refactor(compression-layer): merge PassthroughCompressionLayer into CompressionLayer
